+++ src/org/loklak/api/server/AppsServlet.java
@@ -38,6 +38,7 @@ import org.json.JSONObject;
 import org.loklak.data.DAO;
 import org.loklak.http.RemoteAccess;
 import org.loklak.server.FileHandler;
+import org.loklak.server.Query;
 
 public class AppsServlet extends HttpServlet {
 
@@ -50,7 +51,7 @@ public class AppsServlet extends HttpServlet {
     
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        RemoteAccess.Post post = RemoteAccess.evaluate(request);
+        Query post = RemoteAccess.evaluate(request);
 
         String callback = post.get("callback", "");
         boolean jsonp = callback != null && callback.length() > 0;
