+++ src/org/loklak/api/server/StatusServlet.java
@@ -40,6 +40,7 @@ import org.loklak.data.DAO;
 import org.loklak.http.RemoteAccess;
 import org.loklak.objects.QueryEntry;
 import org.loklak.server.FileHandler;
+import org.loklak.server.Query;
 import org.loklak.tools.OS;
 
 public class StatusServlet extends HttpServlet {
@@ -53,7 +54,7 @@ public class StatusServlet extends HttpServlet {
     
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        RemoteAccess.Post post = RemoteAccess.evaluate(request);
+        Query post = RemoteAccess.evaluate(request);
         
         String callback = post.get("callback", "");
         boolean jsonp = callback != null && callback.length() > 0;
