+++ src/org/loklak/api/server/UserServlet.java
@@ -21,6 +21,7 @@ package org.loklak.api.server;
 
 import java.io.IOException;
 import java.io.PrintWriter;
+
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
@@ -30,6 +31,7 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.loklak.harvester.TwitterAPI;
 import org.loklak.http.RemoteAccess;
+import org.loklak.server.Query;
 
 import twitter4j.TwitterException;
 
@@ -44,7 +46,7 @@ public class UserServlet extends HttpServlet {
     
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        RemoteAccess.Post post = RemoteAccess.evaluate(request);
+        Query post = RemoteAccess.evaluate(request);
      
         // manage DoS
         if (post.isDoS_blackout()) {response.sendError(503, "your request frequency is too high"); return;}
