+++ src/org/loklak/data/IndexFactory.java
@@ -21,7 +21,6 @@ package org.loklak.data;
 
 import java.io.IOException;
 import java.util.Collection;
-import java.util.List;
 import java.util.Set;
 
 import org.json.JSONObject;
